+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## 0.5.8
 
+* `Common`:
+    * New extension `Iterable#firstNotNull`
+* `Coroutines`
+    * New extension `Flow#firstNotNull`
+    * New extensions `CoroutineContext#LinkedSupervisorJob`, `CoroutineScope#LinkedSupervisorJob` and
+      `CoroutineScope#LinkedSupervisorScope`
+
 ## 0.5.7
 
 * `Pagination`